LayoutTests/imported/w3c/web-platform-tests/eventsource/format-mime-bogus-expected.txt

CONSOLE MESSAGE: EventSource's response has a MIME type ("x bogus") that is not "text/event-stream". Aborting the connection.

PASS EventSource: bogus MIME type
